Consolidate workflow jobs; test library with 1.46 MSRV and CLI with 1.54 (#79)
* cargo: Update to count-zeroes 0.2.1

Needed for Rust 2018 compatibility.

* workflows: Consolidate PR workflow into main

They're nearly identical, so combine them into one workflow with
conditionals.

* workflows: Consolidate main jobs using strategy matrix

They run nearly-identical code.  Avoid code duplication using a matrix
and a couple conditionals.

* workflows: Test on separate MSRVs for lib/CLI

Test the library with Rust 1.46 and the CLI with Rust 1.54, in addition
to stable.  The CLI already uses std::time::Duration::ZERO, which
requires 1.53+, and we're about to switch to clap 3 which requires
1.54+.

See discussion in https://github.com/cecton/gptman/pull/77.

* README: Add MSRV badge

* workflows: Run clippy against CLI

https://github.com/cecton/gptman/pull/81 forgot to enable clippy runs
against the CLI now that we no longer build it by default.