Update workflow actions to latest (#121)
Fix various deprecation warnings. The `actions-rs` and
`actions/*release*` workflows are unmaintained; replace the former with
manual Rust commands and the latter with a substitute.

Use generic content type for uploaded artifacts. This avoids the need to
run the release step once per artifact, and the content types don't seem
to be attached to the final download anyway.

Also drop monthly run of Rust workflow. We already have the `cargo
update` PR to ensure there's a monthly CI run. Removing the timer from
the Rust job allows it to be enabled in forks for pre-PR testing without
also running it periodically there.